Move source file under pkg. Reorder the import.
