+++ pkg/commands/agnhost_test.go
