+++ pkg/commands/netcat.go
