+++ pkg/commands/netcat_test.go
