+++ pkg/entities/kubernetes/error.go
