+++ pkg/matrix/analyzer.go
