+++ pkg/matrix/result.go
