+++ pkg/matrix/truthtable.go
