+++ tests/conntrack_test.go
@@ -6,14 +6,14 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/k8sbykeshed/k8s-service-validator/entities"
-	"github.com/k8sbykeshed/k8s-service-validator/entities/kubernetes"
-	"github.com/k8sbykeshed/k8s-service-validator/matrix"
 	"go.uber.org/zap"
+	v1 "k8s.io/api/core/v1"
 	"sigs.k8s.io/e2e-framework/pkg/envconf"
 	"sigs.k8s.io/e2e-framework/pkg/features"
 
-	v1 "k8s.io/api/core/v1"
+	"github.com/k8sbykeshed/k8s-service-validator/pkg/entities"
+	"github.com/k8sbykeshed/k8s-service-validator/pkg/entities/kubernetes"
+	"github.com/k8sbykeshed/k8s-service-validator/pkg/matrix"
 )
 
 var (
